+++ packages/visual-editor/src/components/puck/cards/index.ts
@@ -1,0 +1,5 @@
+export {
+  ExampleRepeatableItemComponent,
+  type ExampleRepeatableItemProps,
+} from "./ExampleRepeatableItem.tsx";
+export { ProductCard, type ProductCardProps } from "./ProductCard.tsx";
